requirements.txt
# ============================================
# WEB FRAMEWORK (FastAPI instead of Flask)
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================
# CORS
# ============================================
# FastAPI has built-in CORS (no separate package needed)

# ============================================
# DATABASE (PostgreSQL + Async)
# ============================================
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# ============================================
# WEBSOCKET (Real-time)
# ============================================
websockets==12.0
python-socketio==5.10.0

# ============================================
# LLM - Groq
# ============================================
groq==0.11.0

# ============================================
# VALIDATION
# ============================================
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================
# ENVIRONMENT
# ============================================
python-dotenv==1.0.1

# ============================================
# DATE/TIME
# ============================================
python-dateutil==2.8.2

# ============================================
# HTTP REQUESTS
# ============================================
httpx==0.27.0
requests==2.32.0

# ============================================
# ASYNC SUPPORT
# ============================================
aiofiles==23.2.1
anyio==4.2.0

# ============================================
# TESTING
# ============================================
pytest==8.3.0
pytest-asyncio==0.24.0

# ============================================
# DEV TOOLS (Optional)
# ============================================
black==24.8.0
flake8==7.1.0
